A2A/dem2/pyproject.toml
[project]
name = "dem2"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "a2a-sdk>=0.3.25",
    "crewai>=1.11.1",
    "crewai-tools>=1.11.1",
    "ddgs>=9.11.4",
    "duckduckgo-search>=8.1.1",
    "google-adk>=1.14.1",
    "langchain-community>=0.4.1",
    "langchain-google-genai>=4.2.1",
    "plotly>=6.6.0",
    "uvicorn>=0.42.0",
]
